requirements.txt
torch
torchvision
Pillow
numpy
tqdm
pyyaml
matplotlib
python-Levenshtein